Create Optimization
This was the first attempt at optimizing the neural network. The idea was to drop five variables and see which combination would yield the best mean absolute error. However, we found that some mae values were lower than what we got using all 18 variables. This should not be the case if the network were optimized. 
This code also takes quite a while to run since there are 6,188 possible combinations of five variables.