Create Variable Optimization
In this code, we are now trying to do the variable optimization except using the correlation between variables. We create a heatmap of the correlated variables, which tells us which variables are "highly" correlated. In this attempt, I try to drop every variable associated with one variable and then do the same for a different variable. I also am using automatic epoch stopping, or the network will stop training once the mae value starts to increase again. This should fix some things about the network; for instance, I tried dropping pairs of correlated variables (variables with correlations above 0.65), but sometimes the mae value increased significantly when dropping only one of those values.